+++ weatherEffects/effects/StarsWeatherEffect.js
@@ -71,7 +71,7 @@ export class StarsWeatherEffect extends AbstractWeatherEffect {
       scale: 1,
       direction: 180,
 			density: Math.round(100 * p) / 100,
-			tint: "#FFFFFF",
+			tint: "#bee8ee",
 			period: Math.round(100 * this.CONFIG.lifetime.min / p) / 100
     }
   }